#3 new apps academic info+ director app

+++ TeachingManagement/AcademicInfoApp/views.py
@@ -1,8 +1,8 @@
 from django.shortcuts import render
 from django.contrib import messages
 from django.shortcuts import render, redirect,get_object_or_404
-from .models import Field,Section,School,Degree,Courses,TypeProfessor,Language,Year
-from .forms import FieldForm,SectionForm,SchoolForm,DegreeForm,CoursesForm,TypeProfessorForm,LanguageForm,YearForm
+from .models import Field,Section,School,Degree,Course,TypeProfessor,Language,Year
+from .forms import FieldForm,SectionForm,SchoolForm,DegreeForm,CourseForm,TypeProfessorForm,LanguageForm,YearForm
 
 # Create your views here.
 
@@ -13,25 +13,22 @@ def field_list(request):
     fields = Field.objects.all()
     deleting = None
 
-    if request.method == "POST":
+    # FINAL DELETE
+    if request.method == "POST" and 'confirm_delete' in request.POST:
+        field_id = request.POST.get('confirm_delete') # id passed in url
+        try:
+            field = Field.objects.get(pk=field_id)
+            field_name = field.NameField  # Store the name for the message
+            field.delete()
+            messages.success(request, f"El camp {field_name} s'ha eliminat correctament.")
+            return redirect('field_list') 
+        except Field.DoesNotExist:
+            messages.error(request, "Error: El camp no existeix.")
 
-       # FINAL DELETE
-        if 'confirm_delete' in request.POST:
-            field_id = request.POST.get('confirm_delete') # id passed in url
-            try:
-                field = Field.objects.get(pk=field_id)
-                field_name = field.NameField  # Store the name for the message
-                field.delete()
-                messages.success(request, f"El camp {field_name} s'ha eliminat correctament.")
-                return redirect('field_list') 
-            except Field.DoesNotExist:
-                messages.error(request, "Error: El camp no existeix.")
-                print("Error: El camp amb ID", field_id, "no existeix.")  
-    
     # ACTION OF INITIAL DELETE
     if 'confirm_delete' in request.GET:
         field_id = request.GET.get('confirm_delete')
-        deleting = get_object_or_404(Field, pk=field_id)  # Get id field to delete
+        deleting = field_id  # Get id field to delete
 
     return render(request, 'field_list_actions.html', {
         'fields': fields,
@@ -71,25 +68,23 @@ def field_create_edit(request, idField=None):
 def section_list(request):
     sections = Section.objects.all()
     deleting = None
-   
-    if request.method == "POST":
 
-       # FINAL DELETE
-        if 'confirm_delete' in request.POST:
-            section_id = request.POST.get('confirm_delete') # id passed in url
-            try:
-                section = Section.objects.get(pk=section_id)
-                section_name = section.NameSection  # Store the name for the message
-                section.delete()
-                messages.success(request, f"La secció {section_name} s'ha eliminat correctament.")
-                return redirect('section_list') 
-            except Section.DoesNotExist:
-                messages.error(request, "Error: La secció no existeix.")
+    # FINAL DELETE
+    if request.method == "POST" and 'confirm_delete' in request.POST:
+        section_id = request.POST.get('confirm_delete') # id passed in url
+        try:
+            section = Section.objects.get(pk=section_id)
+            section_name = section.NameSection  # Store the name for the message
+            section.delete()
+            messages.success(request, f"La secció {section_name} s'ha eliminat correctament.")
+            return redirect('section_list') 
+        except Section.DoesNotExist:
+            messages.error(request, "Error: La secció no existeix.")
     
     # ACTION OF INITIAL DELETE
     if 'confirm_delete' in request.GET:
         section_id = request.GET.get('confirm_delete')
-        deleting = get_object_or_404(Section, pk=section_id)  # Get id section to delete
+        deleting = section_id # Get id section to delete
 
     return render(request, 'section_list_actions.html', {
         'sections': sections,
@@ -124,27 +119,28 @@ def sections_create_edit(request,idSection=None):
 
 
 ### SCHOOLS ###
+
+## Schools list to manage - listing and actions of edit, delete and add schools
 def school_list(request):
     schools = School.objects.all()
     deleting = None
    
-    if request.method == "POST":
-       # FINAL DELETE
-        if 'confirm_delete' in request.POST:
-            school_id = request.POST.get('confirm_delete') # id passed in url
-            try:
-                school = School.objects.get(pk=school_id)
-                school_name = school.NameSchool  # Store the name for the message
-                school.delete()
-                messages.success(request, f"L'escola {school_name} s'ha eliminat correctament.")
-                return redirect('school_list') 
-            except School.DoesNotExist:
-                messages.error(request, "Error: La secció no existeix.")
+    if request.method == "POST" and 'confirm_delete' in request.POST:       
+        # FINAL DELETE
+        school_id = request.POST.get('confirm_delete') # id passed in url
+        try:
+            school = School.objects.get(pk=school_id)
+            school_name = school.NameSchool  # Store the name for the message
+            school.delete()
+            messages.success(request, f"L'escola {school_name} s'ha eliminat correctament.")
+            return redirect('school_list') 
+        except School.DoesNotExist:
+            messages.error(request, "Error: L'escola no existeix.")
     
     # ACTION OF INITIAL DELETE
     if 'confirm_delete' in request.GET:
         school_id = request.GET.get('confirm_delete')
-        deleting = get_object_or_404(School, pk=school_id)  # Get id school to delete
+        deleting = school_id  # Get id school to delete
     
     return render(request, 'school_list_actions.html', {
         'schools': schools,
@@ -177,171 +173,171 @@ def school_create_edit(request,idSchool=None):
     return render(request, 'school_form.html', {'form': form})
     
 ### DEGREE ###
-def degree_crud(request):
+
+## Degree list to manage - listing and actions of edit, delete and add degree
+def degree_list(request):
     degrees = Degree.objects.all()
-    form = DegreeForm()  # Initialize the form
     deleting = None
-
-    if request.method == "POST":
-        # FINAL DELETE
-        if 'confirm_delete' in request.POST:
-            degree_id = request.POST.get('confirm_delete')
-            print("Intentant eliminar el grau amb ID:", degree_id)
-            try:
-                degree = Degree.objects.get(pk=degree_id)
-                degree_name = degree.NameDegree  # Store the name for the message
-                degree.delete()
-                messages.success(request, f"El grau {degree_name} s'ha eliminat correctament.")
-                return redirect('degree_crud')
-            except Degree.DoesNotExist:
-                messages.error(request, "Error: El grau no existeix.")
-                print("Error: El grau amb ID", degree_id, "no existeix.")
-
-        # UPDATE
-        if 'idDegree' in request.POST:
-            degree_id = request.POST.get('idDegree')
-            print("Intentant actualitzar el grau amb ID:", degree_id)
-            degree = get_object_or_404(Degree, pk=degree_id)  # Check if degree exists
-            form = DegreeForm(request.POST, instance=degree)
-            if form.is_valid():  # Check if form is valid
-                form.save()
-                messages.success(request, f"El grau {degree.NameDegree} s'ha actualitzat correctament.")
-                return redirect('degree_crud')
-            else:
-                print("Errors en el formulari d'actualització:", form.errors)
-                messages.error(request, "Error en actualitzar el grau. Si us plau, revisa els camps.")
-
-        # CREATE
-        else:
-            form = DegreeForm(request.POST)  # Initialize form with POST data
-            print("Intentant crear un nou grau:", request.POST)
-            if form.is_valid():  # Check if form is valid
-                degree = form.save()
-                messages.success(request, f"El grau {degree.NameDegree} s'ha creat correctament.")
-                return redirect('degree_crud')
-            else:
-                print("Errors en el formulari de creació:", form.errors)
-                messages.error(request, "Error en crear el grau. Si us plau, revisa els camps.")
-
-    # UPDATE FORM
-    if 'edit' in request.GET:
-        degree_id = request.GET.get('edit')
-        print("Intentant editar el grau amb ID:", degree_id)
-        degree = get_object_or_404(Degree, pk=degree_id)
-        form = DegreeForm(instance=degree)
-
+   
+    if request.method == "POST" and 'confirm_delete' in request.POST:
+       # FINAL DELETE
+        degree_id = request.POST.get('confirm_delete') # id passed in url
+        try:
+            degree = Degree.objects.get(pk=degree_id)
+            degree_name = degree.NameDegree  # Store the name for the message
+            degree.delete()
+            messages.success(request, f"La Titulació {degree_name} s'ha eliminat correctament.")
+            return redirect('degree_list') 
+        except Degree.DoesNotExist:
+            messages.error(request, "Error: La Titulació no existeix.")
+    
     # ACTION OF INITIAL DELETE
     if 'confirm_delete' in request.GET:
         degree_id = request.GET.get('confirm_delete')
-        deleting = get_object_or_404(Degree, pk=degree_id)  # Get degree to delete
-
-    return render(request, 'degree_crud.html', {
-        'form': form,
+        deleting = degree_id  # Only pass the ID for now
+    
+    return render(request, 'degree_list_actions.html', {
         'degrees': degrees,
         'deleting': deleting,
     })
 
-### COURSES ###
-def course_crud(request):
-    courses = Courses.objects.all()
-    form = CoursesForm()
-    deleting = None
-
-    if request.method == "POST":
-        # Handle delete confirmation
-        if 'confirm_delete' in request.POST:
-            course_id = request.POST.get('confirm_delete')
-            course = get_object_or_404(Courses, pk=course_id)
-            course_name = course.NameCourse
-            course.delete()
-            messages.success(request, f"El curs {course_name} s'ha eliminat correctament.")
-            return redirect('courses_crud')
-
-        # Handle update
-        if 'idCourse' in request.POST:
-            course_id = request.POST.get('idCourse')
-            course = get_object_or_404(Courses, pk=course_id)
-            form = CoursesForm(request.POST, instance=course)
+#Function to create or edit a degree - depends if is passed a idegree 
+def degree_create_edit(request,idDegree=None):
+    if idDegree:
+        # If idDegree is passed, we are editing an existing degree
+        degree = get_object_or_404(Degree, pk=idDegree)
+        if request.method == 'POST':
+            form = DegreeForm(request.POST, instance=degree)
             if form.is_valid():
                 form.save()
-                messages.success(request, f"El curs {course.NameCourse} s'ha actualitzat correctament.")
-                return redirect('courses_crud')
-
-        # Handle create
+                messages.success(request, f"La Titulació {degree.NameDegree} s\'ha actualitzat correctament.")
+                return redirect('degree_list')
         else:
-            form = CoursesForm(request.POST)
+            form = DegreeForm(instance=degree)
+    else:
+        # No idDegree, we are creating a new degree
+        if request.method == 'POST':
+            form = DegreeForm(request.POST)
             if form.is_valid():
-                course = form.save()
-                messages.success(request, f"El curs {course.NameCourse} s'ha creat correctament.")
-                return redirect('courses_crud')
+                new_degree = form.save()
+                messages.success(request, f"La Titulació {new_degree.NameDegree} s\'ha afegit correctament.")
+                return redirect('degree_list')
+        else:
+            form = DegreeForm()
+    return render(request, 'degree_form.html', {'form': form})
 
-    # Handle edit
-    if 'edit' in request.GET:
-        course_id = request.GET.get('edit')
-        course = get_object_or_404(Courses, pk=course_id)
-        form = CoursesForm(instance=course)
 
-    # Handle initial delete confirmation
+### COURSES ###
+
+## Course list to manage - listing and actions of edit, delete and add course
+def course_list(request):
+    courses = Course.objects.all()
+    deleting = None
+   
+    if request.method == "POST" and 'confirm_delete' in request.POST:
+       # FINAL DELETE
+        course_id = request.POST.get('confirm_delete') # id passed in url
+        try:
+            course = Course.objects.get(pk=course_id)
+            course_name = course.NameCourse  # Store the name for the message
+            course.delete()
+            messages.success(request, f"El Curs {course_name} s'ha eliminat correctament.")
+            return redirect('course_list') 
+        except Course.DoesNotExist:
+            messages.error(request, "Error: El Curs no existeix.")
+    
+    # ACTION OF INITIAL DELETE
     if 'confirm_delete' in request.GET:
         course_id = request.GET.get('confirm_delete')
-        deleting = get_object_or_404(Courses, pk=course_id)
+        deleting = course_id  # Only pass the ID for now
 
-    return render(request, 'courses_crud.html', {
-        'form': form,
+    return render(request, 'course_list_actions.html', {
         'courses': courses,
         'deleting': deleting,
     })
 
-### TYPE PROFESSOR ###
-def type_professor_crud(request):
-    type_professors = TypeProfessor.objects.all()
-    form = TypeProfessorForm()
-    deleting = None
 
-    if request.method == "POST":
-        # Handle delete confirmation
-        if 'confirm_delete' in request.POST:
-            type_professor_id = request.POST.get('confirm_delete')
-            type_professor = get_object_or_404(TypeProfessor, pk=type_professor_id)
-            type_professor_name = type_professor.Name
-            type_professor.delete()
-            messages.success(request, f"El tipus de professor {type_professor_name} s'ha eliminat correctament.")
-            return redirect('type_professor_crud')
-
-        # Handle update
-        if 'idTypeProfessor' in request.POST:
-            type_professor_id = request.POST.get('idTypeProfessor')
-            type_professor = get_object_or_404(TypeProfessor, pk=type_professor_id)
-            form = TypeProfessorForm(request.POST, instance=type_professor)
+#Function to create or edit a course - depends if is passed a idCourse 
+def course_create_edit(request,idCourse=None):
+    if idCourse:
+        # If idCourse is passed, we are editing an existing courses
+        course = get_object_or_404(Course, pk=idCourse)
+        if request.method == 'POST':
+            form = CourseForm(request.POST, instance=course)
             if form.is_valid():
                 form.save()
-                messages.success(request, f"El tipus de professor {type_professor.Name} s'ha actualitzat correctament.")
-                return redirect('type_professor_crud')
-
-        # Handle create
+                messages.success(request, f"El curs {course.NameCourse} s\'ha actualitzat correctament.")
+                return redirect('course_list')
         else:
-            form = TypeProfessorForm(request.POST)
+            form = CourseForm(instance=course)
+    else:
+        # No idCourse, we are creating a new course
+        if request.method == 'POST':
+            form = CourseForm(request.POST)
             if form.is_valid():
-                type_professor = form.save()
-                messages.success(request, f"El tipus de professor {type_professor.Name} s'ha creat correctament.")
-                return redirect('type_professor_crud')
+                new_course = form.save()
+                messages.success(request, f"El curs {new_course.NameCourse} s\'ha afegit correctament.")
+                return redirect('course_list')
+        else:
+            form = CourseForm()
+    return render(request, 'course_form.html', {'form': form})
 
-    # Handle edit
-    if 'edit' in request.GET:
-        type_professor_id = request.GET.get('edit')
-        type_professor = get_object_or_404(TypeProfessor, pk=type_professor_id)
-        form = TypeProfessorForm(instance=type_professor)
+### TYPE PROFESSOR ###
 
-    # Handle initial delete confirmation
+## typeprofessor list to manage - listing and actions of edit, delete and add typeprofessor
+def typeprofessor_list(request):
+    typeprofessors = TypeProfessor.objects.all()
+    deleting = None
+   
+    if request.method == "POST" and 'confirm_delete' in request.POST:
+       # FINAL DELETE
+        typeprofessor_id = request.POST.get('confirm_delete') # id passed in url
+        try:
+            typeprofessor = TypeProfessor.objects.get(pk=typeprofessor_id)
+            typeprofessor_name = typeprofessor.NameContract  # Store the name for the message
+            typeprofessor.delete()
+            messages.success(request, f"El Tipus de Professor {typeprofessor_name} s'ha eliminat correctament.")
+            return redirect('typeprofessor_list') 
+        except TypeProfessor.DoesNotExist:
+            messages.error(request, "Error: El Tipus de professor no existeix.")
+    
+    # ACTION OF INITIAL DELETE
     if 'confirm_delete' in request.GET:
-        type_professor_id = request.GET.get('confirm_delete')
-        deleting = get_object_or_404(TypeProfessor, pk=type_professor_id)
+        typeprofessor_id = request.GET.get('confirm_delete')
+        deleting = typeprofessor_id  # Only pass the ID for now
 
-    return render(request, 'type_professor_crud.html', {
-        'form': form,
-        'type_professors': type_professors,
+    return render(request, 'typeprofessor_list_actions.html', {
+        'type_professors': typeprofessors,
         'deleting': deleting,
     })
+
+
+#Function to create or edit a typeprofessor - depends if is passed a idTypeProfessor 
+def typeprofessor_create_edit(request,idTypeProfessor=None):
+    if idTypeProfessor:
+        # If idTypeProfessor is passed, we are editing an existing TypeProfessor
+        typeprofessor = get_object_or_404(TypeProfessor, pk=idTypeProfessor)
+        if request.method == 'POST':
+            form = TypeProfessorForm(request.POST, instance=typeprofessor)
+            if form.is_valid():
+                form.save()
+                messages.success(request, f"El curs {typeprofessor.NameContract} s\'ha actualitzat correctament.")
+                return redirect('typeprofessor_list')
+        else:
+            form = TypeProfessorForm(instance=typeprofessor)
+    else:
+        # No idTypeProfessor, we are creating a new typeprofessor
+        if request.method == 'POST':
+            form = TypeProfessorForm(request.POST)
+            if form.is_valid():
+                new_typeprofessor = form.save()
+                messages.success(request, f"El curs {new_typeprofessor.NameContract} s\'ha afegit correctament.")
+                return redirect('typeprofessor_list')
+        else:
+            form = TypeProfessorForm()
+    return render(request, 'typeprofessor_form.html', {'form': form})
+
+########################################################################################################
 
 
 ### LANGUAGES ###
